refactor functions;
